devenv: prep runtime directory paths
